LayoutTests/fast/mediastream/stopping-image-capture-heap-collect-crash-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: grabFrame is stopped

Harness Error (FAIL), message = Unhandled rejection: grabFrame is stopped

PASS grabFrame while GC
